feat: complete app router migration (Chunk 8)
