+++ src/app/me/dashboard/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { Layout } from "@/components/Me/Layout";
+
+export default function DashboardPage() {
+  return (
+    <Layout>
+      <h1>TODO</h1>
+    </Layout>
+  );
+}
