+++ src/components/AdminOnly/AdminOnly.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useSession } from "next-auth/react";
 import { FC } from "react";
