+++ src/components/Header/Header.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Disclosure } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
 import Image from "next/image";
@@ -5,7 +7,7 @@ import Link from "next/link";
 import { BsTelegram, BsGithub, BsLinkedin, BsTwitter } from "react-icons/bs";
 import { MdWork } from "react-icons/md";
 import classNames from "classnames";
-import { useRouter } from "next/compat/router";
+import { usePathname } from "next/navigation";
 
 const navigation = [
   // { name: "Contatti", href: "#", current: false },
@@ -46,7 +48,7 @@ const navigation = [
 ];
 
 export const Header = () => {
-  const router = useRouter();
+  const pathname = usePathname();
   return (
     <Disclosure as="nav">
       {({ open }) => (
@@ -86,13 +88,13 @@ export const Header = () => {
                         target={item.external ? "_blank" : undefined}
                         rel="noopener noreferrer"
                         className={classNames(
-                          item.current && router?.pathname === item.href
+                          item.current && pathname === item.href
                             ? "bg-slate-700 text-white dark:bg-black dark:bg-opacity-40 dark:text-slate-300"
                             : "text-slate-600 hover:bg-slate-200 dark:text-slate-300 dark:hover:bg-black dark:hover:bg-opacity-20",
                           "flex items-center gap-2 rounded-md px-3 py-2 text-sm font-medium",
                         )}
                         aria-current={
-                          item.current && router?.pathname === item.href
+                          item.current && pathname === item.href
                             ? "page"
                             : undefined
                         }
@@ -115,14 +117,14 @@ export const Header = () => {
                   key={item.name}
                   href={item.href}
                   aria-current={
-                     item.current && router?.pathname === item.href
+                     item.current && pathname === item.href
                        ? "page"
                        : undefined
                   }
                 >
                   <span
                     className={classNames(
-                      item.current && router?.pathname === item.href
+                      item.current && pathname === item.href
                         ? "bg-slate-700 text-white dark:bg-black dark:bg-opacity-40 dark:text-slate-300"
                         : "text-slate-600 hover:bg-slate-200 dark:text-slate-300 dark:hover:bg-black dark:hover:bg-opacity-20",
                       "flex cursor-pointer items-center gap-2 rounded-md px-3 py-2 text-base font-medium",
