--- src/pages/_document.tsx
@@ -1,13 +1,0 @@
-import { Html, Head, Main, NextScript } from "next/document";
-
-export default function Document() {
-  return (
-    <Html>
-      <Head />
-      <body className="dark:bg-slate-900 dark:text-slate-300" lang="it">
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  );
-}
