--- src/pages/me/dashboard/index.tsx
@@ -1,12 +1,0 @@
-import { Layout } from "@/components/Me/Layout";
-import { NextPage } from "next";
-
-const DashboardPage: NextPage = () => {
-  return (
-    <Layout>
-      <h1>TODO</h1>
-    </Layout>
-  );
-};
-
-export default DashboardPage;
