--- src/pages/me/proposals/new.tsx
@@ -1,12 +1,0 @@
-import { ProposalForm } from "@/components/Me/ProposalForm";
-import { Layout } from "@/components/Me/Layout";
-
-const NewEventPage = () => {
-  return (
-    <Layout title="Nuova proposta">
-      <ProposalForm />
-    </Layout>
-  );
-};
-
-export default NewEventPage;
